few further tweaks and making multi-tenancy opt-in
