+++ services/api/.env
@@ -31,6 +31,9 @@ API_URL=http://localhost:2300
 UPLOADS_STORE=local
 UPLOADS_GCS_BUCKET=bedrock-staging-uploads
 
+# Multi-Tenancy
+DEFAULT_ORGANIZATION_NAME=
+
 # Sentry Error Tracking
 SENTRY_DSN=
